tests for AnswerItem and AnswerList
